- 150 warnings fixed.

+++ server/src/common/item.c
@@ -137,7 +137,7 @@ char * describe_attack(const object *const op, int newline)
 char * query_weight(object *op)
 {
     static char buf[10];
-    int         i   = op->nrof ? op->nrof *op->weight : op->weight + op->carrying;
+    unsigned int i = op->nrof ? op->nrof *op->weight : (unsigned int) (op->weight + op->carrying);
 
     if (op->weight < 0)
         return "      ";
